Ignore items with missing asset metadata
When finding the asset-level metadata that all items have in common, we should not drop a field just because one item happens to be missing that field, if all the other items had the same value for it.

For example, some items have `eo:bands`, some don't: https://github.com/sat-utils/sat-api/issues/229. When computing `eo:bands` for the DataArray, just ignore the items that are missing it.