Handle some errors as nodata
Pass error patterns as `errors_as_nodata`, which if raised, will be printed as a warning, and whatever they were trying to load will be returned as an empty array.

The main thing I'm unsure about is the `warnings.warn`. In brief testing, it didn't reliably show up, and I'm not sure whether it would show up in distributed worker logs. However, I was also testing a URL that only intermittently failed, so I can't be sure the error was actually happening.

I'll likely merge this now, since it's still a significant improvement, but we should come back to the larger logging story.

Closes #12