Update examples to use `us-west-2` (#49)
Now that Coiled supports `us-west-2`, we can use that!

This also adds a script to build software environments in parallel in multiple regions. By creating envs in multiple regions up front, the first person won't have to wait for an image build when they spin up a cluster in one of the demos.

Also added `jupyterlab-system-monitor` for fun.