Notes to self about why `show` is so laggy:
It seems to be something broken with BatchedSend (what a surprise!).
_maybe_ something to do with visualizing something that's been persisted
(Futures that depend on other Futures)? Locally at least, commenting out
the `persist()` seems to make things work smoothly.

But the main problem we see is messages getting stuck in the BatchedSend
and not getting sent. If you look at `client.scheduler_comm` when things
should be computing but aren't, you'll usually see that there are
messages in the buffer.

In these cases, `scheduler_comm.waker` is not set. This seems wrong;
calling send should set the event. Unless the deadline is set.

But also, `waker._waiters` is empty, as though nothing is waiting on the
event either. I suppose this would be the case while `background_send`
is awaiting the `comm.write`. Or there's some broken thread-safety
(https://github.com/dask/distributed/issues/5552).

Another thing I've noticed is that `scheduler_comm.next_deadline -
scheduler_comm.loop.time()` goes very, very negative (up to hundreds of
seconds) while there are things in the buffer. As in it's hundreds of
seconds behind when it should have sent.
One question is what happens the deadline for `waker.wait` has already
passed. The "avoid sprious wakeups" logic is assuming that so long as
any deadline is set, the coroutine is going to wake up soon anyway. This
would only be the case if that timeout actually works. It's possible
that a `comm.write` takes longer than the interval, and by the time we
come back around to `waker.wait`, the deadline has already passed. If
the wait returns immediately in that case, then we're fine. But if
tornado schedules the callback into the ether, then we'll never wake up.
Though this still doesn't line up with the empty `_waiters` on the
event.

I guess the last explanation is just that `await comm.write` is really,
really, really slow. If it takes hundreds of seconds to do the write
(including serialization), then of course the next batch wouldn't be
sent yet. Websocket comms do have some ridiculous copying. And maybe
because the binder machines and network are so wimpy, it's just extra
slow there? Feel like this doesn't quite add up though.