WIP: switch to PDM
PDM is capable of locking in a few mins what Poetry seems to take 10s of mins (hours?) to solve with the numpy-pandas black hole: https://github.com/python-poetry/poetry/issues/5121

It also supports overriding dependencies, which Poetry refuses and is critical in basic development (especially if working with a fork of distributed).

TODO: update docs, binder, etc. to use PDM. Figure out CI. Al that. This is just an experiment.