avoid mutating `spec.vrt_params`
It's not a good idea in dask to mutate inputs. This object could be shared with other tasks.

Also, I think we just need to check `ds.crs is None`—if `ds.crs` is a CRS, it will always have a `to_epsg` method. (Though that could fail, so I've added handling for that case too.)